requirements.txt
asgiref==3.7.2
Django==5.0.3
django-debug-toolbar==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
Faker==24.2.0
# faker-file[images]==0.17.11
gunicorn==21.2.0
MarkupSafe==2.1.5
packaging==24.0
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
Werkzeug==3.0.1
